requirements.txt
torch~=2.6.0
einops~=0.8.1
numpy~=1.26.4
h5py~=3.13.0
matplotlib~=3.9.4
opencv-python~=4.11.0.86
pillow~=11.1.0
tifffile~=2024.8.30
